Adjust delays
There's no need for delays between matrix jobs - they have
similar write bandwidth, and in fact it is more representative
if the write workload is continuous. So remove it.

On the other hand, introduce a long delay after the prefill job
so that the disk can destage all the writes from its cache to
media.